+++ tests/testthat/test-plot-residuals-vs-simulated.R
@@ -1,0 +1,122 @@
+# data to be used ---------------------------------------
+
+context("plotResidualsVsSimulated")
+
+skip_if_not_installed("vdiffr")
+skip_if(getRversion() < "4.1")
+
+# load the simulation
+sim <- loadTestSimulation("MinimalModel")
+simResults <- importResultsFromCSV(
+  simulation = sim,
+  filePaths = getTestDataFilePath("Stevens_2012_placebo_indiv_results.csv")
+)
+
+# import observed data (will return a list of DataSet objects)
+dataSet <- loadDataSetsFromExcel(
+  xlsFilePath = getTestDataFilePath("CompiledDataSetStevens2012.xlsx"),
+  importerConfiguration = loadDataImporterConfiguration(getTestDataFilePath("ImporterConfiguration.xml"))
+)
+
+# create a new instance and add datasets
+myCombDat <- DataCombined$new()
+myCombDat$addDataSets(dataSet)
+myCombDat$addSimulationResults(
+  simResults,
+  quantitiesOrPaths = c(
+    "Organism|Lumen|Stomach|Metformin|Gastric retention",
+    "Organism|Lumen|Stomach|Metformin|Gastric retention distal",
+    "Organism|Lumen|Stomach|Metformin|Gastric retention proximal"
+  )
+)
+
+myCombDat$setGroups(
+  names = c(
+    "Organism|Lumen|Stomach|Metformin|Gastric retention",
+    "Organism|Lumen|Stomach|Metformin|Gastric retention distal",
+    "Organism|Lumen|Stomach|Metformin|Gastric retention proximal",
+    "Stevens_2012_placebo.Placebo_total",
+    "Stevens_2012_placebo.Placebo_distal",
+    "Stevens_2012_placebo.Placebo_proximal"
+  ),
+  groups = c("Solid total", "Solid distal", "Solid proximal", "Solid total", "Solid distal", "Solid proximal")
+)
+
+test_that("It creates default plots as expected", {
+  set.seed(123)
+  vdiffr::expect_doppelganger(
+    title = "default plot",
+    fig = plotResidualsVsSimulated(myCombDat)
+  )
+})
+
+
+test_that("It doesn't work with log scale for Y-axis", {
+  myPlotConfiguration <- DefaultPlotConfiguration$new()
+  myPlotConfiguration$yAxisScale <- tlf::Scaling$log
+  expect_error(
+    plotResidualsVsSimulated(myCombDat, myPlotConfiguration),
+    messages$logScaleNotAllowed()
+  )
+})
+
+
+test_that("It respects custom plot configuration", {
+  myPlotConfiguration <- DefaultPlotConfiguration$new()
+  myPlotConfiguration$yUnit <- ospUnits$Fraction$`%`
+  myPlotConfiguration$title <- "My Plot Title"
+  myPlotConfiguration$subtitle <- "My Plot Subtitle"
+  myPlotConfiguration$caption <- "My Sources"
+  myPlotConfiguration$legendTitle <- "My legendary title"
+  myPlotConfiguration$legendTitleColor <- "red"
+  myPlotConfiguration$legendTitleSize <- 16
+  myPlotConfiguration$legendKeysColor <- "brown"
+  myPlotConfiguration$pointsSize <- 2.5
+  myPlotConfiguration$legendPosition <- tlf::LegendPositions$outsideRight
+  myPlotConfiguration$pointsColor <- tlf::ColorMaps$default
+  myPlotConfiguration$linesLinetype <- names(tlf::Linetypes)
+
+  set.seed(123)
+  vdiffr::expect_doppelganger(
+    title = "customized plot",
+    fig = plotResidualsVsSimulated(
+      myCombDat,
+      myPlotConfiguration
+    )
+  )
+
+  # Since these were not specified by the user, they should not be updated
+  # after plotting function is done with it.
+  expect_null(myPlotConfiguration$xLabel)
+  expect_null(myPlotConfiguration$yLabel)
+})
+
+
+# edge cases ------------------------
+
+test_that("It returns `NULL` when `DataCombined` is empty", {
+  myCombDat <- DataCombined$new()
+
+  expect_null(suppressWarnings(plotResidualsVsSimulated(myCombDat)))
+  expect_warning(
+    plotResidualsVsSimulated(myCombDat),
+    messages$plottingWithEmptyDataCombined()
+  )
+})
+
+
+test_that("It returns `NULL` when `DataCombined`doesn't have any pairable datasets", {
+  dataSet1 <- DataSet$new(name = "Dataset1")
+  dataSet1$setValues(1, 1)
+  dataSet1$yDimension <- ospDimensions$`Concentration (molar)`
+  dataSet1$molWeight <- 1
+
+  myCombDat <- DataCombined$new()
+  myCombDat$addDataSets(dataSet1)
+
+  expect_null(suppressWarnings(plotResidualsVsSimulated(myCombDat)))
+  expect_warning(
+    plotResidualsVsSimulated(myCombDat),
+    messages$plottingWithNoPairedDatasets()
+  )
+})