+++ NEWS.md
@@ -2,9 +2,10 @@
 
 * Adds new visualization functions:
 
-    - `plotObservedVsSimulated()` for observed versus predicted scatter plot.
-    - `plotResidualsVsTime()` for time versus residuals scatter plot.
-
+    - `plotObservedVsSimulated()` for observed versus predicted data scatter plot.
+    - `plotResidualsVsTime()` for time versus residuals data scatter plot.
+    - `plotResidualsVsSimulated()` for predicted versus residuals data scatter plot.
+    
 # ospsuite 11.0.123
 
 ## New features